refactor(app): extract provider connect auth views
